frameworks/fm_intentkit/pyproject.toml
[project]
name = "fm-intentkit"
version = "0.1.0"
description = "IntentKit framework implementation for BBH benchmarking"
requires-python = ">=3.12"
dependencies = [
    "intentkit @ git+https://github.com/crestalnetwork/intentkit.git",
    "requests>=2.25.0",
    "datasets==3.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
]
setup = "ready"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]